重构代码结构
